Add a `quoted` function; remove auto-quoting of strings with `.`
The previous behaviour was to quote identifiers that contain a
dot(`.`). Doing this prevents us from being able to access fields
using `<table>.<field>` since it takes the entire identifier to be
just the field name.

This commit adds a `quoted` function which instead allows you to
explicitly quote an identifier if it has dots or hyphens.